feat: move discussion related components to discussions folder
