+++ app/frontend/src/features/communities/CommunityPage/index.ts
@@ -1,3 +1,2 @@
 export { default, useCommunityPageStyles } from "./CommunityPage";
-export { default as DiscussionCard } from "./DiscussionCard";
 export { default as SectionTitle } from "./SectionTitle";
