+++ app/frontend/src/features/communities/discussions/DiscussionCard.tsx
