+++ app/frontend/src/features/communities/discussions/index.ts
@@ -1,2 +1,3 @@
 export { default as DiscussionPage } from "./DiscussionPage";
 export { default as DiscussionsListPage } from "./DiscussionsListPage";
+export { default as DiscussionsSection } from "./DiscussionsSection";
